Add option to skip health check for some repositories
Some large git repos (e.g. the Linux Kernel) take an excessively long
time to fsck, and the resulting failure messages pollute the Gitea
system notices with noise.

Add the field SKIP_REPOS to the [cron.repo_health_check] section of
app.ini which is a list of repos for which to skip the health check.

Implements: https://github.com/go-gitea/gitea/issues/1712
Signed-off-by: Allen Wild <allenwild93@gmail.com>