version script: fix for v1.6
This time upstream managed to forget to update the version in main.go,
so the codeversion check failed. Just drop that and always use git
describe output if the branch doesn't have "master" in the name.

Also fix the sed to only s/-/+/ after the tag name rather than the first
'-' (since v1.6.0-rc1 has a dash in it already)