admin: make garbage collection and fsck run in the background
git gc takes way too long to be a synchronous action.
This is a reimplementation of 71d2654d64a48498f9ba3248246ada89c8c3386d
after it was lost during the latest merge where upstream moved the
GitGcRepos function to a new file.

Also that apparently made fsck run in the foreground (blocking page
reload) so give it a similar treatment, though GitFsck still needs to
return an error for use with cron.